Move Theme Toggle to the Toolbar
This replaces theme theme <Select /> with a Light/Dark toggle button.

This implementation feels slightly hacky, and there is slight regression (unless
we present a select tag, we currently lose the ability to have a “system” option
that adapts based on system theme.

Also adds icon hovering to bottom toolbar.